add tools module
