+++ pyhodl/config.py
@@ -21,8 +21,6 @@
 import os
 from enum import Enum
 
-from pyhodl.data.coins import Coin, CoinsNamesTable
-
 
 class RunMode(Enum):
     """ Run as ... """
@@ -74,9 +72,6 @@ COINS_DATABASE = os.path.join(
     RAW_DATA_FOLDER,
     "coins.json"
 )
-FIAT_COINS = [Coin("USD"), Coin("EUR")]  # supported fiat coins
-DEFAULT_FIAT = "USD"
-CRYPTO_COINS = CoinsNamesTable(COINS_DATABASE).get_coins()
 
 DEFAULT_PATHS = {
     RunMode.STATS: DATA_FOLDER,
@@ -94,14 +89,3 @@ def get_coin_historical_data_file(currency):
     """
 
     return os.path.join(HISTORICAL_DATA_FOLDER, currency.lower() + ".json")
-
-
-def is_crypto(coin):
-    """
-    :param coin: str or Coin
-        Coin to check
-    :return: bool
-        True iff coin is among crypto supported
-    """
-
-    return coin not in FIAT_COINS
